Introduce invocation logger functionality
In complex programs and frameworks object interactions and invocations
can be hard to understand. Similarly, debugging in complex scenarios can
be aggravated by too few information about the involved components and
objects.
This change introduces functionality to intercept and log invocations of
public methods on chosen objects or classes of objects. It works by
replacing those methods with wrappers that log passed in arguments to as
well as return values from the original method. In order to log all
invocations of methods on objects of a certain class, a subclass is
created -- a process that is simplified by provided APIs.