Do not fix logging level
Until now we always used the debug log level for logging our
invocations. This approach is suboptimal because first it ties us
unnecessarily to an interpretation of when to log data and second, it
kind of introduces a dependency towards the logging package (that is not
entirely true as one could create arbitrary objects that sport a 'debug'
method, however, not many objects used for outputting data do so
natively).
This change adjusts the logging mechanism and with it the interface we
require a logger object to provide. With it, we assume a message can be
logged just by "calling" (or invoking) the logger object. This way we no
longer require a 'debug' method in the logger object and we also allow
clients to provide arbitrary objects for printing data. The downside (if
one can say so) is that in order to use the functionality in conjunction
with Python's 'logging' module one has to pass in a lambda function
invoking the log method for the desired log level.