Add support for handling meta classes correctly
Our approach of creating a subclass of the class to log and mixing in a
meta class works fine for "normal" classes. It fails, however, once the
target class already contains a meta class. Python quits such problems
with an exception:
> TypeError: metaclass conflict: the metaclass of a derived class must
> be a (non-strict) subclass of the metaclasses of all its bases

The problem can be illustrated as follows:
  Foo <- MetaFoo
   |
   v
  Bar <- InvocationLogger

Here, 'Bar' is the class that's method invocations are to be logged. It
is derived from 'Foo' which already has a meta class applied, 'MetaFoo'.
Once we apply the 'InvocationLogger' meta class we run into the
TypeError. The problem is that Python expects 'InvocationLogger' in the
example to be a subclass of 'MetaFoo'.
This change adjusts the subclass creation in the Logged() function
accordingly.