Do not use custom attributes for parameter passing
Until now we used __meta_original_classname__ for conveying the original
class' name to the logging infrastructure and __meta_logger__ to pass
the logger object to it. These attributes were defined within the
wrapped class, unnecessarily polluting the namespace.
This change removes these attributes and  instead passes the values of
interest to the meta class directly (and hands them through to the
destination by means of method parameters).