introduce object based wrapping
Until now it was only possible to wrap entire classes to have logging of
method invocations. Although it is always possible to create a new class
for each object to create, a separate mechanism for wrapping objects can
be preferable, for instance, if the class' constructor cannot be invoked
but an instance is already available.
This change enhances the module to provide functionality for wrapping
objects as opposed to only entire classes.

Note that the names for the two approach ('Logged' for class based and
'LoggedObj' for object based wrapping) are now slightly non-orthogonal.
The reason for this naming is to emphasize the preference of the class
based approach.